pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for market research agent"
authors = [
    {name = "Michael Hua",email = "michael.huayl@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic-ai-slim[openai,tavily] (>=0.4.4,<0.5.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
show-fixes = true
target-version = "py312"
line-length = 80

[tool.ruff.lint]
extend-select = ["N", "UP", "PL", "Q"]
ignore = ["E501"]
